Restore const enums
