+++ common/api-review/storage.api.md
@@ -12,6 +12,8 @@ import { Metadata } from '@firebase/storage-types/exp';
 import { StorageObserver } from '@firebase/storage-types/exp';
 import { StorageReference } from '@firebase/storage-types/exp';
 import { StorageService } from '@firebase/storage-types/exp';
+import { TaskEvent } from '@firebase/storage-types/exp';
+import { TaskState } from '@firebase/storage-types/exp';
 import { UploadResult } from '@firebase/storage-types/exp';
 import { UploadTask } from '@firebase/storage-types/exp';
 
@@ -64,25 +66,9 @@ export const StringFormat: {
     DATA_URL: string;
 };
 
-// @public
-export type TaskEvent = string;
+export { TaskEvent }
 
-// @public
-export const TaskEvent: {
-    STATE_CHANGED: string;
-};
-
-// @public
-export type TaskState = string;
-
-// @public
-export const TaskState: {
-    RUNNING: string;
-    PAUSED: string;
-    SUCCESS: string;
-    CANCELED: string;
-    ERROR: string;
-};
+export { TaskState }
 
 // @public
 export function updateMetadata(ref: StorageReference, metadata: Partial<Metadata>): Promise<Metadata>;
