+++ packages/storage/src/implementation/constants.ts
@@ -19,38 +19,6 @@
  */
 
 /**
- * @public
- * Error codes that can be attached to `FirebaseStorageError`s.
- */
-export const enum StorageErrorCode {
-  // Shared between all platforms
-  UNKNOWN = 'unknown',
-  OBJECT_NOT_FOUND = 'object-not-found',
-  BUCKET_NOT_FOUND = 'bucket-not-found',
-  PROJECT_NOT_FOUND = 'project-not-found',
-  QUOTA_EXCEEDED = 'quota-exceeded',
-  UNAUTHENTICATED = 'unauthenticated',
-  UNAUTHORIZED = 'unauthorized',
-  RETRY_LIMIT_EXCEEDED = 'retry-limit-exceeded',
-  INVALID_CHECKSUM = 'invalid-checksum',
-  CANCELED = 'canceled',
-  // JS specific
-  INVALID_EVENT_NAME = 'invalid-event-name',
-  INVALID_URL = 'invalid-url',
-  INVALID_DEFAULT_BUCKET = 'invalid-default-bucket',
-  NO_DEFAULT_BUCKET = 'no-default-bucket',
-  CANNOT_SLICE_BLOB = 'cannot-slice-blob',
-  SERVER_FILE_WRONG_SIZE = 'server-file-wrong-size',
-  NO_DOWNLOAD_URL = 'no-download-url',
-  INVALID_ARGUMENT = 'invalid-argument',
-  INVALID_ARGUMENT_COUNT = 'invalid-argument-count',
-  APP_DELETED = 'app-deleted',
-  INVALID_ROOT_OPERATION = 'invalid-root-operation',
-  INVALID_FORMAT = 'invalid-format',
-  INTERNAL_ERROR = 'internal-error',
-  UNSUPPORTED_ENVIRONMENT = 'unsupported-environment'
-}
-/**
  * Domain name for firebase storage.
  */
 export const DEFAULT_HOST = 'firebasestorage.googleapis.com';
