+++ packages/storage/src/implementation/taskenums.ts
@@ -21,13 +21,11 @@
 
 /**
  * An event that is triggered on a task.
- * @public
  */
 export type TaskEvent = string;
 
 /**
  * An event that is triggered on a task.
- * @public
  */
 export const TaskEvent = {
   /**
@@ -60,13 +58,11 @@ export const enum InternalTaskState {
 
 /**
  * Represents the current state of a running upload.
- * @public
  */
 export type TaskState = string;
 
 /**
  * Represents the current state of a running upload.
- * @public
  */
 export const TaskState = {
   /** The task is currently transferring data. */
@@ -84,3 +80,25 @@ export const TaskState = {
   /** The task failed with an error. */
   ERROR: 'error'
 };
+
+export function taskStateFromInternalTaskState(
+  state: InternalTaskState
+): TaskState {
+  switch (state) {
+    case InternalTaskState.RUNNING:
+    case InternalTaskState.PAUSING:
+    case InternalTaskState.CANCELING:
+      return TaskState.RUNNING;
+    case InternalTaskState.PAUSED:
+      return TaskState.PAUSED;
+    case InternalTaskState.SUCCESS:
+      return TaskState.SUCCESS;
+    case InternalTaskState.CANCELED:
+      return TaskState.CANCELED;
+    case InternalTaskState.ERROR:
+      return TaskState.ERROR;
+    default:
+      // TODO(andysoto): assert(false);
+      return TaskState.ERROR;
+  }
+}
